deduplicat format
