+++ apps/back/app/controllers/torrent_controller.ts
@@ -5,6 +5,7 @@ import { inject } from '@adonisjs/core'
 import { join } from 'node:path'
 import app from '@adonisjs/core/services/app'
 import { HttpContext } from '@adonisjs/core/http'
+import { formatBytes } from '../utils/format.js'
 
 interface DeleteMovieResult {
   success: boolean
@@ -100,7 +101,7 @@ export default class TorrentController {
         success: true,
         message: result.message,
         spaceFreed: result.spaceFreed,
-        spaceFreedFormatted: this.formatBytes(result.spaceFreed),
+        spaceFreedFormatted: formatBytes(result.spaceFreed),
       })
     } else {
       return response.internalServerError({
@@ -121,7 +122,7 @@ export default class TorrentController {
         message: result.message,
         moviesDeleted: result.moviesDeleted,
         spaceFreed: result.spaceFreed,
-        spaceFreedFormatted: this.formatBytes(result.spaceFreed),
+        spaceFreedFormatted: formatBytes(result.spaceFreed),
         errors: result.errors,
         errorMessages: result.errorMessages,
       })
@@ -142,15 +143,5 @@ export default class TorrentController {
       message: `An unexpected error occurred while ${operation}`,
       error: error instanceof Error ? error.message : 'Unknown error',
     })
-  }
-
-  private formatBytes(bytes: number): string {
-    if (bytes === 0) return '0 B'
-
-    const k = 1024
-    const sizes = ['B', 'KB', 'MB', 'GB', 'TB']
-    const i = Math.floor(Math.log(bytes) / Math.log(k))
-
-    return `${Number.parseFloat((bytes / Math.pow(k, i)).toFixed(2))} ${sizes[i]}`
   }
 }
